third_party/blink/web_tests/fast/forms/calendar-picker/month-picker-mouse-operations-expected.txt

Tests if month picker mouse operations work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-01"
PASS highlightedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
Check that hovering over an entry highlights it.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "1999-12"
PASS highlightedDayCells() is "1999-12-26,1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-01"
PASS highlightedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
Check that mouse click sets the value.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-01"
PASS highlightedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS document.getElementById("month").value is "2000-01"
Check that clicking the month popup button opens the month popup.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS popupWindow.global.picker.monthPopupView.isVisible is true
PASS highlightedMonthButton() is "2000-01"
Check that hovering over a month button highlights it.
PASS highlightedMonthButton() is "2000-10"
Check that month wheel scrolls the list in the month popup.
PASS checkYearListViewScrollOffset() > 0 is true
PASS checkYearListViewScrollOffset() < 0 is true
Check that clicking the month popup sets the month.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS currentMonth() is "2000-06"
Check that clicking the month popup button opens the month popup.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS popupWindow.global.picker.monthPopupView.isVisible is true
PASS highlightedMonthButton() is "2000-06"
Check that hovering over a month button highlights it.
PASS highlightedMonthButton() is "2000-10"
Check that month wheel scrolls the list in the month popup.
PASS checkYearListViewScrollOffset() > 0 is true
PASS checkYearListViewScrollOffset() < 0 is true
Check that clicking an year list cell opens it.
PASS checkYearListViewScrollOffset() > 0 is true
PASS highlightedMonthButton() is "1999-01"
Check that clicking the month popup sets the month.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS currentMonth() is "1999-06"
PASS successfullyParsed is true

TEST COMPLETE
